api: initial implementation of /extensions node
Xapian usage is dropped with this commit in favour of Elasticsearch.
Actually there is no proper support of Xapian for DRF or Django exists,
but only generic python bindings. Support for Xapian in Haystack is very
poor. Support of Elasticsearch in Haystack is almost dead also.
On the other hand django-elasticsearch-dsl is developed actively for now.