api: initial implementation of /extensions node
This is backport of the wip branch commit that added initial Elasticsearch
implementation that were later replaced by OpenSearch and then merged to
the master branch.

Original message as is:

Xapian usage is dropped with this commit in favour of Elasticsearch.
Actually there is no proper support of Xapian for DRF or Django exists,
but only generic python bindings. Support for Xapian in Haystack is very
poor. Support of Elasticsearch in Haystack is almost dead also.
On the other hand django-elasticsearch-dsl is developed actively for now.