Add CI jobs for building and pushing Docker image to Quay
As part of migration to OpenShift 4, GNOME infrastructure is switching
from internal Docker registry and OpenShift-based builds to Quay and
GitLab CI. After this MR is merged, a new deployment can be created by
launching a manually triggered push-to-quay job. After the new image is
pushed to Quay, OCP4 will pick up a new image within 20 minutes and
start a new deployment automatically.