Add XAPIAN_VERSION to docker-compose.yml
The default Xapian version was removed from the Dockerfile to the
GitLab CI file, which breaks the setup that is described as
"Running with Docker" in the readme. This commit adds the Xapian
version to the docker-compose.yml file, too.

See: https://gitlab.gnome.org/Infrastructure/extensions-web/-/commit/6fc16d7979bf0cfa192a54008dceccd5318eab5b